pyproject.toml
[project]
name = "netbot"
version = "0.8.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.12",
    "audioop-lts>=0.2.1",
    "dateparser>=1.2.1",
    "humanize>=4.12.0",
    "imapclient>=3.0.1",
    "py-cord>=2.6.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[project.urls]
Documentation = "https://github.com/Local-Connectivity-Lab/netbot?tab=readme-ov-file#netbot"
Issues = "https://github.com/Local-Connectivity-Lab/netbot/issues"
Source = "https://github.com/Local-Connectivity-Lab/netbot"

[tool.coverage.run]
source_pkgs = ["netbot", "tests"]
branch = true
parallel = true

[tool.coverage.paths]
netbot = ["main.py", "netbot", "redmine", "threader"]
tests = ["tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

[tool.hatch]

[tool.hatch.version]
path = "netbot/__about__.py"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["netbot", "threader"]

[project.scripts]
netbot = "netbot.__main__:main"
threader = "threader.__main__:main"
